Enhance reviews, ratings, deletion safety, and file uploads
- Change review StarRating to decimal for precision; update Instructor and Section ratings to decimal.
- Auto-update entity ratings (Course, Section, Instructor, Video) on review create/update/delete.
- Prevent deletion of courses/sections with enrolled students; return conflict error if attempted.
- Add support for course thumbnail uploads via IFormFile and Cloudinary (CourseThumbnail category).
- Update course controller endpoints to accept file uploads.
- Remove unused InstructorHomeScreen feature; minor code cleanup and organization.

+++ Application/DTOs/Review/GetAllReviewsResponse.cs
@@ -4,7 +4,7 @@ namespace Application.DTOs.Review
     {
         public Guid Id { get; set; }
         public Guid StudentId { get; set; }
-        public int StarRating { get; set; }
+        public decimal StarRating { get; set; }
         public string Comment { get; set; } = string.Empty;
         public DateTimeOffset CreatedAt { get; set; }
         public DateTimeOffset UpdatedAt { get; set; }
